+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 * Fix: Decode special characters from json.org's escape list correctly
+* Enhancement: Add `login` command
 
 ### [1.1.0](https://github.com/somechris/hallo-eltern-cli/releases/tag/v1.1.0) (2022-11-15)